requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.0.6
django-ninja==1.1.0
djangorestframework==3.15.1
pydantic==2.7.2
pydantic_core==2.18.3
sqlparse==0.5.0
typing_extensions==4.12.0
